requirements.txt
# Core Python version
# python==3.12.3

# Core logic
numpy
faiss-cpu
langchain_text_splitters==1.0.0
huggingface_hub

# File parsing
pymupdf
python-docx

# Web interface
gradio

# Environment variables
python-dotenv
